requirements.txt
Flask==3.0.0
pandas==2.1.4
openpyxl==3.1.2
Werkzeug==3.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
